push_alert/requirements.txt
flask
google-cloude-firestore
numpy
